requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
farama-notifications==0.0.4
filelock==3.20.0
flappy-bird-gymnasium==0.4.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
gymnasium==1.2.3
h5py==3.15.1
idna==3.11
jinja2==3.1.6
keras==3.13.0
kiwisolver==1.4.9
libclang==18.1.1
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml-dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.4.0
opt-einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.2
pygame==2.6.1
pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pyyaml==6.0.3
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
typing-extensions==4.15.0
urllib3==2.6.2
werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
